refactor(earn): EarnForm to own file
